- refactored some reporting code
